requirements.txt
Django~=4.1.1
djangorestframework~=3.13.1
django-environ~=0.9.0
psycopg2~=2.9.3